+++ src/services/Nanuq/Nanuq.Common/Enums/ServerType.cs
@@ -1,0 +1,8 @@
+namespace Nanuq.Common.Enums;
+
+public enum ServerType
+{
+    Kafka,
+    Redis,
+    RabbitMQ
+}