+++ src/services/Nanuq/Nanuq.Common/Exceptions/EncryptionException.cs
@@ -1,0 +1,8 @@
+namespace Nanuq.Common.Exceptions;
+
+public class EncryptionException : Exception
+{
+    public EncryptionException(string message) : base(message) { }
+
+    public EncryptionException(string message, Exception inner) : base(message, inner) { }
+}