+++ src/services/Nanuq/Nanuq.Common/Requests/AddCredentialRequest.cs
@@ -1,0 +1,9 @@
+namespace Nanuq.Common.Requests;
+
+public record AddCredentialRequest(
+    int ServerId,
+    string ServerType,
+    string? Username,
+    string? Password,
+    string? AdditionalConfig
+);