+++ src/services/Nanuq/Nanuq.Common/Requests/UpdateCredentialRequest.cs
@@ -1,0 +1,8 @@
+namespace Nanuq.Common.Requests;
+
+public record UpdateCredentialRequest(
+    int Id,
+    string? Username,
+    string? Password,
+    string? AdditionalConfig
+);